+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.0.2 - 2023-12-08
+
+### Added
+- Add support for Event plugin Events/Tickets.
+
+### Fixed
+- Fix Neo field exports.
+- Fix export dates not working correctly for non-US-based locales.
+
+## 1.0.1 - 2023-11-05
+
+### Fixed
+- Fix an error installing on Postgres installs.
+- Fix disabled dropdown values in preview, due to selectize.
+
 ## 1.0.0 - 2023-09-19
 
 ### Added